requirements.txt
pandas
Pillow
fastapi
uvicorn
python-multipart
openpyxl
google-generativeai
pdf2image
requests
